requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
pypdf
mistralai==0.4.2